fixed compiler warnings

+++ source/adios2/engine/ssc/SscReader.cpp
@@ -78,8 +78,8 @@ StepStatus SscReader::BeginStep(const StepMode stepMode,
     {
         if (m_MpiMode == "twosided")
         {
-            MPI_Waitall(m_MpiRequests.size(), m_MpiRequests.data(),
-                        MPI_STATUS_IGNORE);
+            MPI_Waitall(static_cast<int>(m_MpiRequests.size()),
+                        m_MpiRequests.data(), MPI_STATUS_IGNORE);
             m_MpiRequests.clear();
         }
         else if (m_MpiMode == "onesidedfencepush")
